+++ src/BlScraper.DependencyInjection/ConfigureBuilder/IMapQuest.cs
@@ -1,6 +1,3 @@
-using BlScraper.DependencyInjection.Builder;
-using BlScraper.DependencyInjection.Builder.Internal;
-
 namespace BlScraper.DependencyInjection.ConfigureBuilder;
 
 /// <summary>
@@ -17,31 +14,4 @@ public interface IMapQuest
     /// Get available quests
     /// </summary>
     IEnumerable<Type> GetAvailableQuests();
-}
-
-internal class MapQuest : IMapQuest
-{
-    private IEnumerable<(Type Quest, Type Data)> _availableQuests;
-
-    public MapQuest(ScrapBuilderConfig assemblyBuilderAdd)
-    {
-        _availableQuests = assemblyBuilderAdd.Assemblies.SelectMany((assembly) =>
-        {
-            return assembly.GetTypes().Where(t => TypeUtils.IsTypeValidQuest(t)).Select((type) =>
-            {
-                var model = new ScrapModelInternal(type);
-                return (model.QuestType, model.DataType);
-            });
-        }).ToList();
-    }
-
-    public IEnumerable<Type> GetAvailableQuests()
-    {
-        return _availableQuests.Select(tuple => tuple.Quest);
-    }
-
-    public IEnumerable<(Type Quest, Type Data)> GetAvailableQuestsAndData()
-    {
-        return _availableQuests;
-    }
 }
